+++ dcpy/connectors/filesystem.py
@@ -6,24 +6,24 @@ from dcpy.connectors.registry import StorageConnector
 
 
 class Connector(StorageConnector):
-    conn_type: str = "drive"
+    conn_type: str = "filesystem"
     path: Path | None = None
 
-    def _path(self, key: str, drive_path: Path | None = None) -> Path:
+    def _path(self, key: str, path_prefix: Path | None = None) -> Path:
         if self.path:
             path = self.path / key
-        elif drive_path:
-            path = drive_path / key
+        elif path_prefix:
+            path = path_prefix / key
         else:
             path = Path(key)
         path.parent.mkdir(parents=True, exist_ok=True)
         return path
 
-    def exists(self, key, drive_path: Path | None = None) -> bool:
-        return self._path(key, drive_path).exists()
+    def exists(self, key, path_prefix: Path | None = None) -> bool:
+        return self._path(key, path_prefix).exists()
 
-    def get_subfolders(self, prefix: str, drive_path: Path | None = None) -> list[str]:
-        path = self._path(prefix, drive_path)
+    def get_subfolders(self, prefix: str, path_prefix: Path | None = None) -> list[str]:
+        path = self._path(prefix, path_prefix)
         return [str(p.relative_to(path)) for p in path.glob("*") if p.is_dir()]
 
     def pull(
@@ -31,10 +31,10 @@ class Connector(StorageConnector):
         key: str,
         destination_path: Path,
         *,
-        drive_path: Path | None = None,
+        path_prefix: Path | None = None,
         **kwargs,
     ) -> dict:
-        source_path = self._path(key, drive_path)
+        source_path = self._path(key, path_prefix)
         if source_path != destination_path:
             shutil.copy(source_path, destination_path)
         return {"path": destination_path}
@@ -44,11 +44,11 @@ class Connector(StorageConnector):
         key: str,
         *,
         filepath: Path,
-        drive_path: Path | None = None,
+        path_prefix: Path | None = None,
         overwrite: bool = False,
         **kwargs,
     ) -> dict:
-        path = self._path(key, drive_path)
+        path = self._path(key, path_prefix)
 
         if path.exists() and not overwrite:
             raise Exception(f"'{path}' already exists.")
