+++ dcpy/lifecycle/connector_registry.py
@@ -2,7 +2,7 @@ from dcpy.configuration import RECIPES_BUCKET
 from dcpy.connectors.edm import recipes, publishing
 from dcpy.connectors.socrata.connector import SocrataConnector
 from dcpy.connectors.esri.arcgis_feature_service import ArcGISFeatureServiceConnector
-from dcpy.connectors import drive, web, s3, ingest_datastore
+from dcpy.connectors import filesystem, web, s3, ingest_datastore
 from dcpy.connectors.registry import (
     ConnectorRegistry,
     Connector,
@@ -22,7 +22,7 @@ def _set_default_connectors():
     connectors.register(connector=ArcGISFeatureServiceConnector())
     connectors.register(connector=web.WebConnector())
     connectors.register(connector=web.WebConnector(), conn_type="api")
-    connectors.register(connector=drive.Connector(), conn_type="local_file")
+    connectors.register(connector=filesystem.Connector(), conn_type="local_file")
     connectors.register(connector=s3.S3Connector(), conn_type="s3")
     connectors.register(
         connector=s3.S3Connector(bucket=RECIPES_BUCKET, prefix="datasets/"),
